core:Power Management: restart operatio..(#865385)
core:Power Management: restart operation fails occasionally

Currently we are performing the following in the Restart command
1) send Stop command to the fence agent
2) wait for 'off' status by sampling the agent status command
3) send Start command to the fence agent
4) wait for 'on' status by sampling the agent status command

There are some cases in which the Start command can fail but the engine
will still try to get the 'on' status and give up after the configured
retries. The reason is that the Start & Stop command are async and we
only get success for the fence script invocation.
In the case that the Agent & Host uses the same power supply the Start
may fail, therefor we had decided that in any case that we fail to Start
the Host after the configurable retries, we will send additional Start
command to cover such a case.

Change-Id: I1e6a515fb246d86778d359f26ce26eae8e52ebc5
Signed-off-by: Eli Mesika <emesika@redhat.com>