core: don't re-migrate vm after migration was canceled (#867347)
https://bugzilla.redhat.com/show_bug.cgi?id=867347

if migration was canceled, vm went back to up on the source,
this is identified as failure in migration, and re-run is called,
causing migration to run again.

this patch remove the vm from the async-running-vms list on
cancel-migration,
that way the vm will not be re-run when migration will be identified as
failed.

Change-Id: I9a8bf38bef90e5dfd0458cc13b8bb704f11d9f58
Signed-off-by: Omer Frenkel <ofrenkel@redhat.com>