core, restapi: Convert NFS version to enum
In order to support optional advanced nfs options, the "auto" version
specifier must be distinct from "not set" in the database.  Instead,
"auto" will be explicit as a string, and "not set" will be null.  This
change requires the core and rest api to propagate non-numeric version
values as well.

This conversion will also help later when minor versions are specified
in the mount options, e.g. "4.1".

Bug-Url: https://bugzilla.redhat.com/871534
Change-Id: I09d82db7c0ec92135cae8935f013487e97b8f357
Signed-off-by: Greg Padgett <gpadgett@redhat.com>