core: adding unlock_entity.sh script(#868638)
This patch adds a script to unlock disk objects in the db.
The script has 2 modes
 1. unlock a VM/Template
 2. unlock a VM/Template and all its associated disks (using the -r
flag)
 3. unlock a specific disk

Only plugged disks are considered as candidates of this command.

The new script can be invoked basically by

unlock_entity.sh -t TYPE -i ID
where
     TYPE is either "vm" or "template" or "disk"
     ID is the entity name in case of VM/Tempalte, UUID is case of
     a disk.
In order to display locked entities use:
unlock_entity.sh -t TYPE -q

     The script supports all regular DB parameters as USER DATABASE
     etc. including the option to run the script remotely.

The user is prompted to confirm that this is a dangerous operation and
he should contact GSS prior to this operation.

The operation is logged as an Alert in the audit_log with all relevant
details.

Change-Id: I3f1712fb945933d92ee9a6b97bc71d4ba3a01597
Signed-off-by: Eli Mesika <emesika@redhat.com>
Bug-Url:https://bugzilla.redhat.com/868638