core: ISO domain permission error shouldn't tirgger SPM selection.
Failure while executing getFloppyList command with
permission error triggers DC reinitialization instead of just logging
the error.

GetFloppyListCommand is a IrsBrokerCommands,
which means that generally an exception thrown by it would
trigger a run of the failure mechanism in IrsBrokerCommand.

This patch changes the default thrown exception from this commands to
be an exception that doesn't trigger failover, so unless we encounter
specific case SPM reselection won't be triggered.

Change-Id: I4e75982d1a88dc36251d2c2e693ffafdc7a7a99e
Bug-Url: https://bugzilla.redhat.com/854975
Signed-off-by: Liron Aravot <laravot@redhat.com>