rest: User can't see its permissions (#868304)
https://bugzilla.redhat.com/868304

When asking for object permissions in the REST API, we call search
query, to get all users.

This patch changes that to only fetch for users relevant for the
permission itself, doing that using GetDbUserById query, and also making
it a user query to allow users to view permissions on their objects.

Change-Id: I6f21f4b8a845d52276dd7d97b5c7069892bb3b77
Signed-off-by: Oved Ourfali <oourfali@redhat.com>