rest: Support filtered capabilities (#868949)
https://bugzilla.redhat.com/868949

Several config parameters aren't exposed to users (as they aren't
relevant). These are used in the capabilities resource.
Changing the generation of the capabilities response to not contain
these values.
Also, exposing the cluster levels, as they may be relevant for users as
well.
Finally, removing the enable local storage test, as it is always enabled
in 3.1.

Change-Id: I549e680226f0d802872915b8a46c9e5a8c322b44
Signed-off-by: Oved Ourfali <oourfali@redhat.com>