core: Enable addDisk with a locked disk (#870754)
In AddDiskCommand - check only if the VM is locked and not any of its
disks.
This Fixes a regression entered in
https://bugzilla.redhat.com/show_bug.cgi?id=865551
in patch http://gerrit.ovirt.org/#/c/8666/ .
The original bug fixed an error message using a different flow than the
original one. The new flow remains and a parameter to
ImagesHandler.PerformImagesChecks() was altered.

Change-Id: Ifb3714b8aa301fd12a2837e00d899b09cc0001f6
Bug-Url: https://bugzilla.redhat.com/870754
Signed-off-by: Vered Volansky <vvolansk@redhat.com>