webadmin: Fix localization issue in edit quota
Edit Quota popup view had some hard coded strings
which are imposible to localize to some languages.

The string origin in the ApplicationConstants so they
could not be formated.

The issue was also relevant to  SubTabQuotaClusterView and
SubTabQuotaStorageView and was fix in those classes as well.

New formated strings added to ApplicationMessages and the hard coded
part was replaced with a message formating.

This fix also handles a second localization issue in the
EditQuotaClusterPopupView. The "vCpus" label was moved from
EditQuotaClusterPopupView.ui.xml to ApplicationConstants to allow
localization.

Change-Id: I73146871d6ca7a8f91a190816df2c2dc684d9f5e
Bug-Url: https://bugzilla.redhat.com/871042
Signed-off-by: Ofri Masad <omasad@redhat.com>