engine: Preparation for removing a global lock on spm calls at engine side
The following patch is preporation for removing a global lock on spm operation at engine side.
In order to do this as a first stage I am removing or reduce an usage of methods which can require
additional synchronization.
Thes patch is removing some usage of getCurrentVdsId() and removed unused commands

Change-Id: I3044261122112b5557953c92dd80782d4edc444e
Signed-off-by: Michael Kublin <mkublin@redhat.com>