packaging: Updated interaction with user on async tasks
* The previous try/finally clause included the initial interaction with
the user, where tasks were found and user was asked whether to continue
(and if user continues, then restart engine in maintenance mode), as well
as some operations that could end with an exception.

If, for example, user chooses not to continue, 'finally' will kick in,
and try to restore the engine configuration, although it didn't change yet.
Additionally, the setting "zombie tasks" timeout could fail and raise an
exception, which would also try to restore both an engine AND "zombie tasks
timeout", which is not set yes, leading to an incorrect setting.

This patch aims to solve these issues by moving try/finally block to the
later code, beginning only with the final loop, when a user has agreed
to it.

* Also, added additional try/finally block to the configureForMaintenance
function, so that in case the creation of the temp configuration fails,
we would still try to cleanup the file if it was created and raise the
exception.

* Also, changed the return value in getRunningTasks and getCompensation
functions from "return None" to empty string "return ''". This improves
the printouts to the user, and instead of printing "None" in tasks,
nothing is printed if no tasks are found.

Change-Id: Ia1f1aaf7ac812d69dc81f3fb8dee5a626bc7634a
Signed-off-by: Alex Lourie <alourie@redhat.com>