core: Don't use PostgreSQL UUID extension
Replace the uuid_generate_v1 function from the uuid-ossp extension with
a custom function that generates UUIDs from a sequence. That way we
don't need the extension any longer.

Note that this is the first step to remove completely the generation of
UUIDs in the database, as they should be generated by the engine.

Change-Id: I819e37224288b202d2126cec475ee1d97f83a925
Bug-Url: https://bugzilla.redhat.com/870056
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>