tools : engine-manage-domains location option no longer valid
In command line parameters for engine-manage-domains the -location (=remote/local)
parameter is an obsolete option but the command gives no error when added to the
command, and has no effect.

This also appears with -action=list, prints that the domain is remote, but all domains
are treated as "remote" so there is no point in printing that the domain is remote.

Change-Id: I3384c6981d38b03d2f2856ff8af69ac9d1c65056
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=871085
Signed-off-by: Ravi Nori <rnori@redhat.com>