engine: cannot unpause VM started in paused mode
Have 2 clusters, in each have one host. In one of
this clusters create a VM. Run it in pause mode,
wait until it starts in the pause mode and tham start it.

The VM will not start.

The problem was in VdsSelector in the check where it
tried to find out, if the specific VM is migrating to
the same host.

This check makes sense for the migrating of the VM,
but not for running it.

Fixed by making the checks aware of if the VM is
migrating or not.

This patch also removes one duplication - one of the
checks has been copy-pasted twice.

There was also a workaround which ignored the situation
that the VM would be migrated to the same host - it is
already implmented in the HostValidator so removed.

Change-Id: I8cca521f79bfb7d5762f2c996200916a10afc36a
Bug-Url: https://bugzilla.redhat.com/872205
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>