core: LSM - rollback on non-running VM
LiveMigrateDiskCommand:
Initiating rollback if the VM stops during migration.
On each step of the migration process a new command
is created (in order to execute the next task),
if the VM is no longer running the previous tasks
should be reverted/cleared.

LiveSnapshotTaskHandler:
On compensation, image's status should be reverted to OK.
Updating status using DestinationImageId since ImageId might
be obsolete - depending on which step the failure occurred
(when taking a snapshot, a new active image is generated).

CreateImagePlaceholderTaskHandler:
Rollback should be DeleteImage.

VmReplicateDiskStartTaskHandler:
No rollback for task.

Change-Id: I7aafc5cd63f961470124c984afb2e7a9d74a0de2
Bug-Url: https://bugzilla.redhat.com/872316
Signed-off-by: Daniel Erez <derez@redhat.com>