utils: ssh: move to member '_' convention
We had several discussion over this.

Convention for members/static/globals is very helpful when dealing with
patch hunk and review, as every variable can be assigned to proper
context without reading the whole code.

It also avoid naming clashes between the different scope level.

I am aware that some people are not used to variable name convention,
however it is much easier to maintain and review patches in this mode.

I am enforcing this convention on bootstrap and ssh sources, and I hope
that others will see the benefit in this method as well.

Change-Id: I1bca20cec1e309105805d2c741c56abf04ab33d5
Signed-off-by: Alon Bar-Lev <alonbl@redhat.com>