engine: re-run mechanism improvements
- RunVmCommandBase#rerun method no longer invokes rerunInternal
method within a different thread because there is no need for a
different thread when it is called from the command itself. the
code of rerunInternal method was moved to rerun method without
creating a new thread to run it instead.
when rerun method is invoked by VdsEventListener, its code should
still run in a different thread so it is called from a new thread
in VdsEventListener#rerun method.

- adjust the rerun mechanism such that the logs & events would
reflect the process better:
* on re-run, log the results of the previous run before checking
whether the maximum rerun attempts were exceeded, so that it would
be logged every time.
* add a missing log call after exceeding the maximum re-run attempts
to log a message that indicates the command failed.
* make sure that when a command that is being rerun is failed due to
can-do-action check, the failure event will appear only once.

Change-Id: I557636a94d74772635102c6a33dc923d91332a3e
Signed-off-by: Arik Hadas <ahadas@redhat.com>