core: Failed move disk should not set illegal disk.
When disk failed to be moved to another storage domain,
the engine set the disk status to Illegal with the destination storage
domain.

The proposed fix, set the disk state to OK, with the original storage
domain if the move task fails.
By doing that, the user can use the VM disks even if the move operation
failed.

As part of the fix, some of the code which updated the destination
storage domain in the disk, moved from the execute part to the
EndSuccesfully section.

Change-Id: I22c83bdd2e7182be49156c20992c494bcb68915a
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>