webadmin: Messages text fixes
In this patch, some fixes were made to some of the messages:

- Network engine is mandatory, it cannot be modified.
- The management Network (engine) is mandatory and cannot be removed.

For the messages above (in which "engine" represents the Management
Network), "engine" was replaced with a placeholder for the Management
Network name ("${NetworkName}"), which should be retreived from the
configuration instead of being hard-coded.

Also, the relevant engine commands that were referring to the messages
above were changed to include also a "parameter value" message (in the
form of "$NetworkName [network-name-from-configuration]") in the
relevant canDoActionMessages list.

- The default gateway should be set only on engine network

The message above was changed to:
The default gateway should be set only on the Management Network

At this point, this message cannot contain the actual management network
name placeholder (like the first two messages mentioned in this commit
comment), as this message is being used by the
NetworkInterfaceValidator, which doesn't seem to support multiple
messages / messages with parameters' placeholders.

- Internal oVirt Engine Error
- oVirt Engine has stopped
- oVirt History Service Started
- oVirt History Service Stopped
- oVirt History Service Error
- Could not connect to oVirt Engine Service, please try to refresh
- the page. If the problem persists contact your System Administrator
- Fencing is disabled due to oVirt Engine service start up sequence.
- Permission denied. oVirt Engine Administrator permission is required.
- Authentication Failed. Error in DNS configuration. Please verify
- the oVirt Engine Host has a valid reverse DNS (PTR) record.
- Authentication Failed. oVirt Engine clock is not synchronized with
  directory services (must be within 5 minutes difference).
- Please verify the clocks are synchronized and restart the service.
- Login failed, if the oVirt Engine Service was restarted please wait
  few seconds and try again, if the problem persists please
- check your oVirt Engine license.
- oVirt Engine has detected that the system clock has been modified

The (unnecessary) "oVirt" string was removed from the messages above.

Change-Id: I6368f77f174ff17b4f78856d01236eb07bae8e2a
Bug-Url: https://bugzilla.redhat.com/873036
Signed-off-by: Einav Cohen <ecohen@redhat.com>