engine: Remove IsValidVDSCommand (Preporation for removing global lock on SPM ops)
The following patch will remove IsValidVDSCommand.
The following command was used in order to check if pool is up and we have spm.
First of all the using of command at canDoaction is wrong, a simple check for status of pool or status
of domain should be enough.
But the most exciting part is that the following command can trigger SPM election and it is done without
any synchronization, so we can easilly have a couple spm elections simenteniously.
The code at IsValidVDSCommand was getCurrentIrsProxyData().getIsValid(), which is actually code at IrsProxyData:
getIrsProxy() != null, where getIrsProxy() is elect SPM if we have not one.

Change-Id: I73c0fa0fd66f2f14132594a0b0450a7ecd7cf166
Signed-off-by: Michael Kublin <mkublin@redhat.com>