bootstrap: detach OVirtUpgrader from VdsInstaller into OVirtNodeUpgrade
Currently we have 4 separate bootstrap methods:
 1. Host.
 2. Node registration.
 3. Node installation.
 4. Node upgrade.

While there is something common between (1-3), node upgrade is totally
different sequence:

 1. Unlike bootstrap it is executing foreign code (not originated at
    engine machine.
 2. Unlike bootstrap it is non-customizable.

Current implementation shares the same VdsInstaller base class among
bootstrap and node upgrade, as these are totally different, we detach
the OVirtUpgrader from VdsInstaller.

VdsInstaller is soon to be retired.

Rename OVirtUpgrader to OVirtNodeUpgrade as it handles only ovirt-node.

This change also introduces the SSHDialog class, which is a class that
can be used to conduct a text dialog with a component at the other end
of the ssh session. It is a key component in the future vdsm-bootstrap
rewrite.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=875529
Change-Id: Iff19fdb9f717d424f23bc5d4e5a8df8fce8a58bf
Signed-off-by: Alon Bar-Lev <alonbl@redhat.com>