webadmin: no compat version for exported vm
Due to a change in the backend side, while fetching an exported vm,
the compatibility version is empty.
Therefore the disks subtab in the export dialog was empty (there is a check
for hot_plug_enabled which depends on compat_version, this subtab is reused also
in the disk main tab).

not showing any disk data, prevents the user from changing disk allocation policy,
destination SD and quota.

Change-Id: If72968b984ad1637e879eea5e64dbc38a58828ed
Signed-off-by: Gilad Chaplik <gchaplik@redhat.com>