core: use String instead of StringBuilder in SysprepHandler (#873742)
SysprepHandler uses StringBuilder as it is a string: it creates new
instances for replacements.
this patch changes StringBuilder to String to make it clear to the user
it is regular string that need to be set.

also added return value to populateSysPrepDomainProperties as it
manipulates the string inside it as well, caused changes to domain, user
and pass in the sysprep file to be lost.

Change-Id: Ie447c30759d19e4d42a722deb550a6233becb490
Bug-Url:https://bugzilla.redhat.com/show_bug.cgi?id=873742
Signed-off-by: Omer Frenkel <ofrenkel@redhat.com>