webadmin: reformat remove disk messages
In addition to reformating the "remove disk" messages in order to ease
their translation to the other locales, the "diskNote" key has been
removed from all the non-English .properties files, as the old values
contain variables-placeholders that don't exist in the source (English)
string anymore.

The "diskNote" updated translations will be pushed in a separate patch.

Change-Id: I20172c8e70659a6fa5f77dec738d1c9025daac70
Bug-Id: https://bugzilla.redhat.com/871222
Signed-off-by: Shahar Havivi <shavivi@redhat.com>