core: Fixing http connection leak at proxy for SPM
The following patch will fix a connection leak which we have
during nullify of IrsProxy.
A proxy is representating spm, each time that spm moved to maintaince or
replaced we change a proxy, but coneection of http client was not closed.
Also there are was some interesting code: set proxy to null and after that
to call Dispose() method, where to shutdown connections when a proxy is not null -
magic.
Also I removed some one lines method, they don't help, and caused for a lot bugs
especilaly at that class

Change-Id: Iec0941108ae66c6174626915a45f02ab0996cfee
Signed-off-by: Michael Kublin <mkublin@redhat.com>