core: errors during preview of diskless snapshots
1. When creating a diskless snapshot, then adding a disk to the vm and
later on try to preview that snapshots an NPE occurs within ovf writer.
The exception occurs because of outdated disk information contained within the
vm - the problem wasn't fixed locally but globally to prevent it from
occuring in other flows as well. VmHandler.updateDisksFromDb() should
clear the current disks of the vm otherwise we might get corrupted disk
information for the vm (by adding the disks information from the DB
while not clearing the present information).

2. When trying to run diskless snapshots exception can occur for various
reason in VmHandler.updateVmInSpm(). When such exception occurs the
operation should be rolled back (in the bug scenario - it wasn't as
there were db updates that were executed with no transaction which led
to partially rollback).

Change-Id: I6e8edc74bc34676f526dfd24d2f89eb60d8acf2e
Bug-Url: https://bugzilla.redhat.com/873595
Signed-off-by: Liron Aravot <laravot@redhat.com>