core: Generate UUIDs from engine
We used to generate UUIDs used as primary keys of bookmarks, tags,
clusters hosts and pools in the database. This patch changes the engine
so that it generates them directly and passes them as parameters to the
stored procedures as needed.

Note that some UUIDs can still be generated using the database function
uuid_generate_v1 during first installation and upgrades, as upgrade
scripts haven't been changed.

Change-Id: Idecd3aa570017b4198243a9d3f9ebfdc7a777195
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>