core: Event log - a method to bypass flow
Event log - a method to bypass flood filtering

Adding a customId field to AuditLogableBase in order to enable
distinguishing between events that have same instance references (dc,
cluster, vm etc) but still should be treated as separate events such
that the anti-flood event mechanism will not mask those events and
display just one per the defined period.
It is the responsibility of the AuditLogableBase instance creator to call
setCustomId with a different id per log message instance in order to
treat those messages as different messages.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=875375
Change-Id: Iff160c4241618a2d0c475e13ea13ddefe8220f06
Signed-off-by: Eli Mesika <emesika@redhat.com>