core:Fix NPE when import VM with clone
The following patch should fix an NPE which occurs when importing a
cloned VM or a cloned VM template, which does not exists in the Data
Center.

This exception is caused when trying to copy each disk, and use
the property WipeAfterDelete from the image in the DB,
since the disk does not exists in the DB, we get an NPE.

The proposed fix is to use getDestinationDiskImage to get the
WipeAfterDelete value, instead use the original image with getImage().

Change-Id: I47b8f82a2bd19320d9953df54f9aa16e4ef9200f
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>
Related to Bug-Url: https://bugzilla.redhat.com/862303