core:Fence Host fails if something ...(#871768)
Fence Host fails if something went wrong in FenceQuietTimeBetweenOperationsInSec window [180seconds]

We have a configuration setting of Quiet Time Between PM Operations,
this should control PM operations invoked by the user.
The intension is to prevent subsequent calls to PM stop/start without
waiting to a stable status.

Adding a condition in the FenceVdsBaseCommand canDoAction validation ,
not to check Quiet Time Between PM Operations if the command is internal
(i.e. was invoked by the system and not by the user)

Change-Id: I2cab006a6e911d1eaad888c5c4e41bb4aca68f03
Signed-off-by: Eli Mesika <emesika@redhat.com>