core:Hosts from another data center ...(#874660)
Hosts from another data center displayed when clicked on Hosts in tree view for cluster names with a single difference.

Search Engine is using currently the ILIKE keyword to enable
case-insensitive searches. Postgres treats the character '_' in a value
when using LIKE/ILIKE teh same way as '?' is used for regular
expressions. (i.e. match any single value)
This patch looks for the character '_' in values passed to the search
engine and proceed then with a backslash such that they will be treated
as a regular character and Search Engine will return the correct result.

Change-Id: I2a5e9946f7795827d1abd50f05e641d90802e1e8
Signed-off-by: Eli Mesika <emesika@redhat.com>