core: NPEs and wrong persisted information when reusing LUNs
Generally, the following issues caused that a created storage domain
on pre-existing luns storage led to NPEs and wrong data shown to the user during execution of GetLunsByVgIdQuery.

This patch fixes the following issues:
1. When removing a domain that used LUN storage and there's a disk used by that LUN -
LUNs volume_group_id still contains the removed domain id which doesn't exist anymore.

2. When extending a domain to use an existing LUN, the domain id isn't
set to that LUN which causes to NPE in GetLunsByVgIdQuery or wrong data
provided to the user.

3. When creating a domain from an existing LUN, the domain id isn't set
to this LUN which might causes to NPE in GetLunsByVgIdQuery or
wrong data provided to the user.

4. When removing LUN disk which isn't used as storage domain (and
therefore - not used anymore), the LUN remained in the DB.

Change-Id: I38a0e3c68cb8bd80c2f78ee5aacfccc8c987a79e
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=875909
Signed-off-by: Liron Aravot <laravot@redhat.com>