core:Do not force fencing proxy to be ...(#876235)
Do not force fencing proxy to be in UP status

Currently when a host is non-responsive and has Power Management
configured, we are looking for a proxy host in the same cluster to carry
out the fencing commands that will actually reboot the problematic host.
The algorithm of finding such a proxy was up to now to find the first
host that is in UP status in the host DC.
This patch modified the algorithm of proxy finding to
1) Look for a Host in the problematic Host DC which is UP
2) If 1) fails, try to use any other Host in the same DC as a proxy
   skipping Hosts is non-operational status with NETWORK_UNREACHABLE reason.

This is done because there is no connection between the Host status and
its ability to carry out proxy commands (which can be done if Host has
network connectivity) although we can assume that a Host that is in UP
status is connected to the network and its chance to succeed in carrying
out the fencing operations is the best.

Change-Id: I437e9b2bb0300f8d71cc8f0c50aaff1b253a63bb
Signed-off-by: Eli Mesika <emesika@redhat.com>