packaging: Fix ownership of deployments and contents
A previous commit (9d5af3) changed the .spec so that the backups
directory is owned by the setup sub package. This is needed, but we also
need to make sure that the deployments and contents directories are
still owned by the backend package, otherwise the directories are not
created and the service will not start.

Change-Id: Ibb4745cfc9898b43a164781fd5866cf2b1fa4f82
Bug-Url: https://bugzilla.redhat.com/876164
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>