core:[RFE]Improve power management proxy ...
This patch implements the following RFE
http://wiki.ovirt.org/Features/Design/DetailedHostPMProxyPreferences

In general, adding support for proxy preferences per Host.
Each Host can define how to look for a proxy when the Host should be
fenced.
The default behaviour is to search first in cluster then in the
data-center. This default is available from the engine-config

Change-Id: Id308bc4136a3cc41f622237b05e49b00f11d9bb8
Signed-off-by: Eli Mesika <emesika@redhat.com>