core: Removed unneeded quartz job scheduling during run vm
The following patch is simple, I remove a scheduled quarzt job which
should update the db.
The main problem which is solved it is a job was scheduled each run vm, so
under load we scheduled a number of jobs equals to number of run vms,
it is similar to DDOS attack on internal quartz scheduler.

Change-Id: I8d2365910f471bf01774d3ff7c7fb49520af5386
Signed-off-by: Michael Kublin <mkublin@redhat.com>