core:End failed live snapshot with warning.
Engine fails to create Live-Snapshot during power off of VM.
This is since the KVM process might be killed in the middle of the
process and the synchronization between the engine VM and the KVM process in the VDSM
are not reliable and could be racy.

The proposed solution is to avoid the rollback of the entire operation,
but instead finish the command with success, and add an appropriate
audit log.
That way will allow the VM to be configured to the new snapshot at
the next run.

Change-Id: I1b7f841d3fa02800fd58590c87d64e31227adfe5
Signed-Off-By: Maor Lipchuk <mlipchuk@redhat.com>
Bug-Url: https://bugzilla.redhat.com/870928