packaging: Don't require Maven packages for build
The .spec that we use has a lot of build requirements for maven plugins.
This was copied copied from the official Fedora package, but they are
not needed as we don't build using mvn-rpmbuild but using plain mvn.
This requirements complicate the build as it requires a lot of
additional packages that then aren't used at all.

Change-Id: I9c924ecc53df483976b2bc5ea1777249c120f0ff
Bug-Url: https://bugzilla.redhat.com/878976
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>