core: skip quota for flows with no storage pool
Quota checks are being made based on a commands "quota dependency" which
is being marked by an enum on the VdcActionType. The problem is that
some commands, e.g. AddDisk may need quota check on regular disk but not
on external floating LUN disk, mainly because this command will not have
a storage pool (floating disk doesn't belong to a data center) thus - no
quota could be calculated on it.

As a way out, instead of breaking a command to two, (one quotable and one
not) an empty list of quota params can be submitted so no quota checks
will be made.

Change-Id: Ie468db3709522653cd93c338564592e2ba439c4a
Signed-off-by: Roy Golan <rgolan@redhat.com>