core: Use .pgpass from home in dev env
We use the .pgpass file from /etc/ovirt-engine in production
environments, but this doesn't exist or is not available in most
development environments, so we should try to use the file from user's
home directory. If we don't do that the create_db_devel.sh script keeps
asking for the database password for ever.

Change-Id: I87e36c169dcf3a0aa1914ed46ccea068231d6b6d
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>