core: remove itests
itest are an old testing effort that use an embedded JBoss 5.1 instance
to run bll tests.
Since the upgrade to JBoss 7.1.1, these tests no longer work and are no
longer supported, so they are in effect "dead code".

This patch removes those tests, the resources associated with them and
several non-test classes that are not used in the production code, but
were only used within the itests.

Change-Id: I28cdcae73619f3c94912811d4855cfb869c3c3ff
Signed-off-by: Allon Mureinik <amureini@redhat.com>