core: Fix NPE on Clone VM from Snapshot
Clone VM from snapshot throw NPE while using MoveOrCopyImageGroup.
This happens since AddVmFromSnapshot creates the task first with the
parameters and only after that it creates the disk in the DB.
The proposed fix is to get the wipeAfterDelete flag from the parameters
in case the destination disk is null.
This fix should be revisited, to change the AddVmFromSnapshot to create
the disk before the creation of task.

Change-Id: I2c18b1ac552f598fc20e74aa67f81dabb915a8c2
Signed-Off-By: Maor Lipchuk <mlipchuk@redhat.com>