engine: Add option to create network for public use
The patch adds an option on Network creation to define the
network for public use. If configured for public use, 'everyone'
will get permissions on the network with NetworkUser role.

The motivation is to allow easy access and management of network
so it could be accessed by the system's users.

In addition, since the Networks entities are now managed as part of
the MLA, the creator of the network will be granted with NetworkAdmin
role for the created network.

Change-Id: I14e046c7a98ab6f9de53bec5b8ccc73ac4ba672f
Signed-off-by: Moti Asayag <masayag@redhat.com>