core: remove unneeded reflection use in SyntaxChecker
Using the built-in valueOf method of Enum instead of fancy, expensive
other ways to try to get the enum value.

The former code did eventually a valueOf check which is case sensitive
and then another EnumUtils method which was case *insensitive*. Since the
inner EnumUtils call would only be read on a match there is no point in repeating
the enum fetching twice, and in a very expensive way when we have a
ready made method for it.

Change-Id: Idd74aad49d292e64dec6f5925ffd186f13b9887c
Signed-off-by: Roy Golan <rgolan@redhat.com>