upgrade from 3.0.7 to SI24.4 fail(#879918)
upgrade from 3.0.7 to SI24.4 fails: ERROR: insert or update on table vm_device violates foreign key constraint fk_vm_device_vm _static

This bug is a result of missing FK between vmt_guid in vm_interface and
vm_guid in vm_static.
This cause junk interface left in vm_interface when a template that
has interfaces is removed.
Since in upgrade from 3.0 to 3.1 we added the vm_device table, there is
an attempt to insert those junk interfaces to vm_device that fails on
the FK constraint between vm_device and vm_static.

This patch fix the bug by:
1) Fixing the upgrade script to cleanup those junk interfaces before an
attempt to insert it to vm_device is done.
2) Adds the missing FK between vm_interface and vm_static to prevent
junk data in future.

Change-Id: Iaf4af7908566b5d16e39ea5167c6d6dffc5884db
Signed-off-by: Eli Mesika <emesika@redhat.com>