engine: Query getdisksvmguid caused postmaster processes to consume constantly 100%cpu.
The following bug occurrs because of the following query being run from
GetVmStatsVDSCommand every running VM in the system. The run is cyclic and runs in infinitive loop.
I don't see reason of that query, update of dynamic data ofor disc is done under lock, in worst case
the dynamic data will be updated with wrong values, in next run which occurring every 4 seconds, the
disk will be updated with correct values.
I think we can live with wrong values for 4 seconds in case of race
The change reduced CPU of machine with DB from 100% to 30%, which is still a high, but better

Change-Id: I9ef4a24908a75ab4fc1f02eed7c97a7125a5db41
Bug-Url: https://bugzilla.redhat.com/879930
Signed-off-by: Michael Kublin <mkublin@redhat.com>