webadmin: fix focus traversal in make template
Fixed the focus traversal in the make template popup dialog:
- the focus is switched between the focusable components in the standard
  order (top to down, left to right)
- the focus traversal also include the "ok" and "cancel" buttons

This mechanism can be extended in the future to fix the focus traversal
in other dialogs. it is especially easy to extend it to other similar
popup dialogs (that subclass AbstractModelBoundPopupWidget) - all
we need is that its widget will override "setTabIndexes" method and
define a custom focus order to its components.

Change-Id: I82233e8d136154e942d1ffce8e6f831ddfbfb3fb
Bug-Url: https://bugzilla.redhat.com/873133
Signed-off-by: Arik Hadas <ahadas@redhat.com>