bootstrap: ovirt-node upgrade: create destination directory
One would expect that there will be an ovirt-node upgrade script,
however there is none, and the /data/updates is vdsm specific path,
which was created by vdsm init script, and then creation was moved to
vdsm-reg service.

vdsm-reg is going to be retired soon, so move the creation of this
directory to the component that sends the iso image.

Better solution would be a node upgrade script that receives the image
via stdin.

Change-Id: I5519bd63383feba9b3bf971543b0d630c24bbca0
Signed-off-by: Alon Bar-Lev <alonbl@redhat.com>