core: adding engine Maintenance and Prepare mode.
Support for that was already merged in http://gerrit.ovirt.org/#/c/8669/10

Solution above assumed that configuration values can be reloaded while
the engine is still running. While this is true when working from API, it
is not working from the engine-config, therefor, we must take the config
values directly from the database and not rely on the config cache.

Change-Id: Iae5d1148c9c2868abd0061700bb9988242dd8faf
Signed-off-by: Eli Mesika <emesika@redhat.com>
Bug-Url:https://bugzilla.redhat.com/868639
Signed-off-by: Eli Mesika <emesika@redhat.com>